+++ addOns/automation/CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## Unreleased
 ### Fixed
 - Bug in handling headers with colons in the values.
+- Use default authentication poll frequency when none specified, if the value is less than one a progress warning occurs.
 
 ## [0.51.0] - 2025-07-17
 ### Added